+++ src/spaces/controlPlaneV2/hooks/useCreateManagedControlPlaneV2GraphQL.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@apollo/client/react';
 import { useCallback } from 'react';
 import { z } from 'zod';
 import { buildMcpV2GraphQLInput } from '../helpers/mcpV2GraphQLInput.ts';
-import { McpV2Input, McpV2InputSchema } from '../schemas/mcpV2Input.schema.ts';
+import { McpV2Input, McpV2InputSchema } from '../../mcp/schemas/mcpV2Input.schema.ts';
 import { CreateManagedControlPlaneV2Mutation } from './useCreateManagedControlPlaneV2Mutation.ts';
 
 export function useCreateManagedControlPlaneV2GraphQL() {
