+++ src/spaces/controlPlaneV2/hooks/useCreateManagedControlPlaneV2Mutation.ts
@@ -1,15 +1,14 @@
-import { graphql } from '../../../types/__generated__/graphql/index.ts';
+import { graphql } from '../../../types/__generated__/graphql';
 
-export const UpdateManagedControlPlaneV2Mutation = graphql(`
-  mutation UpdateManagedControlPlaneV2(
-    $name: String!
+export const CreateManagedControlPlaneV2Mutation = graphql(`
+  mutation CreateManagedControlPlaneV2(
     $namespace: String
     $object: CoreOpenControlPlaneIoV2alpha1ControlPlane_Input!
     $dryRun: Boolean
   ) {
     core_open_control_plane_io {
       v2alpha1 {
-        updateControlPlane(name: $name, namespace: $namespace, object: $object, dryRun: $dryRun) {
+        createControlPlane(namespace: $namespace, object: $object, dryRun: $dryRun) {
           metadata {
             name
             namespace
